+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { ThemeProvider } from 'ustudio-ui/theme';
-import App from './core/App';
+import App from './core/app';
 
 import './index.css';
