--- src/modules/main/Card/index.ts
@@ -1,1 +1,0 @@
-export * from './Card.component';
