--- src/modules/main/Form/index.ts
@@ -1,1 +1,0 @@
-export * from './Form.component';
