+++ src/modules/main/card/card.styles.ts
