+++ src/modules/main/card/card.types.d.ts
