+++ src/modules/main/card/index.ts
@@ -1,0 +1,1 @@
+export * from './card.component';
